extract commons of stop loss
