+++ features/automation/protection/stopLoss/state/stopLossStatus.ts
@@ -1,9 +1,21 @@
+import BigNumber from 'bignumber.js'
+import { IlkData } from 'blockchain/ilks'
 import { getToken } from 'blockchain/tokensMetadata'
+import { collateralPriceAtRatio } from 'blockchain/vault.maths'
 import { Vault } from 'blockchain/vaults'
 import { useAppContext } from 'components/AppContextProvider'
 import { PickCloseStateProps } from 'components/dumb/PickCloseState'
-import { closeVaultOptions } from 'features/automation/common/consts'
-import { getIsEditingStopLoss } from 'features/automation/protection/stopLoss/helpers'
+import {
+  closeVaultOptions,
+  DEFAULT_THRESHOLD_FROM_LOWEST_POSSIBLE_SL_VALUE,
+  MIX_MAX_COL_RATIO_TRIGGER_OFFSET,
+} from 'features/automation/common/consts'
+import { SidebarAutomationStages } from 'features/automation/common/types'
+import {
+  checkIfIsDisabledStopLoss,
+  checkIfIsEditingStopLoss,
+  getStartingSlRatio,
+} from 'features/automation/protection/stopLoss/helpers'
 import {
   STOP_LOSS_FORM_CHANGE,
   StopLossFormChange,
@@ -14,29 +26,75 @@ interface GetStopLossStatusParams {
   stopLossTriggerData: StopLossTriggerData
   stopLossState: StopLossFormChange
   isRemoveForm: boolean
+  isProgressStage: boolean
+  isOwner: boolean
+  isAddForm: boolean
+  maxDebtForSettingStopLoss: boolean
   vault: Vault
+  stage: SidebarAutomationStages
+  ilkData: IlkData
 }
 
 interface StopLossStatus {
   isEditing: boolean
+  isDisabled: boolean
   closePickerConfig: PickCloseStateProps
+  resetData: any
+  executionPrice: BigNumber
 }
 
 export function getStopLossStatus({
   stopLossTriggerData,
   stopLossState,
   isRemoveForm,
+  isProgressStage,
+  isOwner,
+  isAddForm,
+  maxDebtForSettingStopLoss,
   vault,
+  stage,
+  ilkData,
 }: GetStopLossStatusParams): StopLossStatus {
   const { uiChanges } = useAppContext()
 
-  const isEditing = getIsEditingStopLoss({
+  const isEditing = checkIfIsEditingStopLoss({
     isStopLossEnabled: stopLossTriggerData.isStopLossEnabled,
     selectedSLValue: stopLossState.stopLossLevel,
     stopLossLevel: stopLossTriggerData.stopLossLevel,
     collateralActive: stopLossState.collateralActive,
     isToCollateral: stopLossTriggerData.isToCollateral,
     isRemoveForm,
+  })
+  const isDisabled = checkIfIsDisabledStopLoss({
+    isAddForm,
+    isEditing,
+    isOwner,
+    isProgressStage,
+    maxDebtForSettingStopLoss,
+    stage,
+  })
+
+  const sliderMin = ilkData.liquidationRatio.plus(MIX_MAX_COL_RATIO_TRIGGER_OFFSET.div(100))
+  const selectedStopLossCollRatioIfTriggerDoesntExist = sliderMin.plus(
+    DEFAULT_THRESHOLD_FROM_LOWEST_POSSIBLE_SL_VALUE,
+  )
+  const initialSlRatioWhenTriggerDoesntExist = getStartingSlRatio({
+    stopLossLevel: stopLossTriggerData.stopLossLevel,
+    isStopLossEnabled: stopLossTriggerData.isStopLossEnabled,
+    initialStopLossSelected: selectedStopLossCollRatioIfTriggerDoesntExist,
+  })
+    .times(100)
+    .decimalPlaces(0, BigNumber.ROUND_DOWN)
+  const resetData = {
+    stopLossLevel: initialSlRatioWhenTriggerDoesntExist,
+    collateralActive: stopLossTriggerData.isToCollateral,
+    txDetails: {},
+  }
+
+  const executionPrice = collateralPriceAtRatio({
+    colRatio: stopLossState.stopLossLevel.div(100),
+    collateral: vault.lockedCollateral,
+    vaultDebt: vault.debt,
   })
   const closePickerConfig = {
     optionNames: closeVaultOptions,
@@ -53,6 +111,9 @@ export function getStopLossStatus({
 
   return {
     isEditing,
+    isDisabled,
+    resetData,
     closePickerConfig,
+    executionPrice,
   }
 }
